native jit for amd64 and arm64; golang-asm, sse fast paths, w^x-hardened code pages

+++ gopherbuzz/vm/value_test.go
@@ -33,7 +33,7 @@ func TestValueConstructorsAndPredicates(t *testing.T) {
 			name:    "FloatValue",
 			v:       vm.FloatValue(3.14),
 			isFloat: true,
-			kind:    "float",
+			kind:    "double",
 		},
 		{
 			name:   "BoolValue true",
